Refactor usage management: Replace Usage model with MeterReading, update related controllers and routes
- Updated usageController to pivot data from MeterReading instead of Usage.
- Removed Usage model and replaced it with MeterReading model.
- Modified BillingPeriod, FeeType, and FeeConfig models for better data handling.
- Enhanced apartment and invoice models with new fields and relationships.
- Updated routes for usage, fees, and invoices to reflect new structure.
- Improved BillCalculator service for tiered fee calculations.
- Adjusted frontend FeeManager to accommodate new fee structure.

+++ backend/models/Payment.js
@@ -4,8 +4,8 @@ const sequelize = require('../config/database');
 const Payment = sequelize.define('Payment', {
   id: { type: DataTypes.INTEGER, autoIncrement: true, primaryKey: true },
   invoice_id: { type: DataTypes.INTEGER, allowNull: false },
-  amount: { type: DataTypes.FLOAT, allowNull: false },
-  method: { type: DataTypes.STRING, allowNull: false }, // BANK, CASH
+  amount: { type: DataTypes.DECIMAL(15, 2), allowNull: false },
+  method: { type: DataTypes.STRING, allowNull: false },
   transaction_code: { type: DataTypes.STRING },
   status: { type: DataTypes.STRING, defaultValue: 'SUCCESS' },
   paid_at: { type: DataTypes.DATE, defaultValue: DataTypes.NOW }
